+++ examples/formatting-ts/examples/write_headers_auto_discover.example.ts
@@ -1,0 +1,17 @@
+import { format } from '@fast-csv/format';
+
+const csvStream = format({ headers: true, writeHeaders: false });
+
+csvStream.pipe(process.stdout).on('end', process.exit);
+
+csvStream.write({ header1: 'value1a', header2: 'value2a' });
+csvStream.write({ header1: 'value1a', header2: 'value2a' });
+csvStream.write({ header1: 'value1a', header2: 'value2a' });
+csvStream.write({ header1: 'value1a', header2: 'value2a' });
+csvStream.end();
+
+// Output:
+// value1a,value2a
+// value1a,value2a
+// value1a,value2a
+// value1a,value2a